Modified structure to dissociate routes from App component
