+++ client/src/tests/NavBar.test.js
@@ -4,5 +4,6 @@ import NavBar from '../NavBar';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<NavBar />, div);
+  const wrapper = shallow(<NavBar />);
+  expect(toJson(wrapper)).toMatchSnapshot();
 });
